refactor: use HttpResult instead of Result<T, AppError>
This commit updates the error handling to use `HttpResult` instead of `Result<T, AppError>`. `HttpResult` is a type alias for `Result<T, AppError>` that is used to differentiate between errors that should be returned to the client and errors that should be logged and handled internally.

This change improves the error handling by making it more explicit when an error should be returned to the client. It also simplifies the error handling code by eliminating the need to check for `AppError` errors and convert them to `HttpResult` errors.